# contracts/user.openapi.yaml
openapi: 3.1.0
info:
  title: Integra — user-service
  version: 2.1.0
  description: |
    Perfil, formação, vínculo institucional e o grafo de seguidores.

    ## A mudança de modelo da v2

    Até a v1 havia **uma** noção de afiliação, obrigatória no cadastro, que ao
    mesmo tempo aparecia no perfil e concedia acesso aos posts internos da
    instituição. Isso misturava duas coisas de naturezas diferentes. A v2 as
    separa:

    **Formação declarada** — o que a pessoa diz. Opcional, sem nenhuma
    verificação, como no LinkedIn: dá para declarar uma faculdade onde nunca se
    estudou. Aparece no perfil e **não concede acesso a nada**.

    **Vínculo institucional** — o que a faculdade confirma. Nasce quando o aluno
    informa o CPF no perfil da faculdade e aquele CPF consta na lista que ela
    cadastrou. É o **único** mecanismo que concede visibilidade de posts
    `institucional` e `curso`.

    ## Duas cardinalidades que não são a mesma

    | | quantidade | efeito |
    | --- | --- | --- |
    | `Formacao` | 0..N, acumula | cosmético; o selo de verificada é permanente |
    | `Vinculo` | **0..1**, se move | concede visibilidade; um por vez |

    Trocar de faculdade encerra o vínculo anterior e abre o novo, mas **as duas
    formações continuam verificadas** — quem se formou na primeira realmente
    estudou lá, e apagar o selo reescreveria o passado. A consequência a não
    esquecer: **formação verificada sem vínculo ativo não concede nada.**
servers:
...
paths:
...
  /users/me:
    get:
      tags: [perfil]
      summary: Perfil do usuário autenticado
      operationId: getMe
      responses:
        '200':
          description: Perfil próprio, com contato, formações e vínculo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Perfil' }
        '401': { $ref: '#/components/responses/NaoAutenticado' }

    patch:
      tags: [perfil]
      summary: Atualiza o próprio perfil
      operationId: updateMe
      description: |
        Campos omitidos ficam inalterados.

        **Não editáveis aqui:** `email` e `cpf`. O e-mail é credencial e pertence
        ao auth-service. O CPF é a chave que liga a conta às matrículas das
        faculdades — deixá-lo editável permitiria assumir a matrícula de outra
        pessoa trocando o próprio CPF pelo dela.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AtualizarPerfilRequest' }
      responses:
        '200':
          description: Perfil atualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Perfil' }
        '401': { $ref: '#/components/responses/NaoAutenticado' }
        '409':
          description: Username já em uso por outra conta
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { code: username_ja_existe, message: 'Username já existe' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /universidades:
    get:
      tags: [instituições]
      summary: Lista universidades
      operationId: listUniversidades
      description: |
        **Sem autenticação** — alimenta o combobox de formação na tela de
        cadastro, antes de existir conta.
      security: []
      parameters:
        - { name: q, in: query, schema: { type: string }, description: Filtro por nome ou sigla. }
      responses:
        '200':
          description: Universidades
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Universidade' }
  /universidades/{universidadeId}:
    get:
      tags: [instituições]
      summary: Perfil público da universidade
      operationId: getUniversidade
      description: |
        A tela que o aluno alcança pela busca. `temVinculo` diz se o leitor já
        está vinculado — é o que decide se o menu oferece "inserir CPF" ou
        "encerrar vínculo".
      parameters:
        - { name: universidadeId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Perfil da instituição
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PerfilDeUniversidade' }
        '401': { $ref: '#/components/responses/NaoAutenticado' }
        '404':
          description: Universidade não encontrada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /universidades/{universidadeId}/cursos:
    get:
      tags: [instituições]
      summary: Cursos de uma universidade
      operationId: listCursos
      description: |
        Os cursos que a própria instituição cadastrou. Alimenta o segundo
        combobox, habilitado depois de escolhida a universidade. **Sem
        autenticação**, pelo mesmo motivo da rota acima.
      security: []
      parameters:
        - { name: universidadeId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Cursos
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Curso' }
        '404':
          description: Universidade não encontrada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /universidades/me/cursos:
    get:
      tags: [instituições]
      summary: Cursos da própria instituição
      operationId: listMeusCursos
      description: Exige `tipo == faculdade` no token.
      responses:
        '200':
          description: Cursos
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Curso' }
        '401': { $ref: '#/components/responses/NaoAutenticado' }
        '403': { $ref: '#/components/responses/SomenteFaculdade' }

    post:
      tags: [instituições]
      summary: Cadastra um curso
      operationId: criarCurso
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nome]
              properties:
                nome: { type: string, maxLength: 200 }
      responses:
        '201':
          description: Curso criado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Curso' }
        '401': { $ref: '#/components/responses/NaoAutenticado' }
        '403': { $ref: '#/components/responses/SomenteFaculdade' }
        '409':
          description: Já existe curso com este nome nesta instituição
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /universidades/me/cursos/{cursoId}:
    delete:
      tags: [instituições]
      summary: Remove um curso
      operationId: removerCurso
      description: |
        Responde `409` se o curso estiver em uso por matrícula, vínculo ou
        formação. Apagar em cascata aqui removeria formação declarada de quem
        não tem relação com a decisão administrativa da instituição.
      parameters:
        - { name: cursoId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Removido }
        '401': { $ref: '#/components/responses/NaoAutenticado' }
        '403': { $ref: '#/components/responses/SomenteFaculdade' }
        '404':
          description: Curso não encontrado nesta instituição
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Curso em uso
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /universidades/me/matriculas:
    get:
      tags: [instituições]
      summary: Alunos cadastrados pela instituição
      operationId: listMatriculas
      description: |
        A lista que a faculdade administra. Cada item diz se já foi reivindicada
        e por quem — é como a secretaria vê quem já entrou no app.
      parameters:
        - { name: cursoId, in: query, schema: { type: string, format: uuid } }
        - name: situacao
          in: query
          schema: { type: string, enum: [todas, pendentes, vinculadas], default: todas }
      responses:
        '200':
          description: Matrículas
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Matricula' }
        '401': { $ref: '#/components/responses/NaoAutenticado' }
        '403': { $ref: '#/components/responses/SomenteFaculdade' }

    post:
      tags: [instituições]
      summary: Cadastra um aluno por CPF
      operationId: criarMatricula
      description: |
        A universidade não vai no corpo: é a do token.

        **A matrícula pode ser criada antes de a conta existir.** É uma lista de
        espera: o vínculo nasce quando alguém com aquele CPF reivindicar. Sem
        isso, a secretaria só poderia cadastrar quem já tivesse baixado o app.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cpf, cursoId]
              properties:
                cpf: { type: string, examples: ['123.456.789-00'] }
                cursoId: { type: string, format: uuid }
      responses:
        '201':
          description: Matrícula criada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Matricula' }
        '401': { $ref: '#/components/responses/NaoAutenticado' }
        '403': { $ref: '#/components/responses/SomenteFaculdade' }
        '409':
          description: Este CPF já está cadastrado nesta instituição
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /universidades/me/matriculas/{matriculaId}:
    delete:
      tags: [instituições]
      summary: Remove o aluno da instituição
      operationId: removerMatricula
      description: |
        Remove a matrícula **e encerra o vínculo ativo**, se existir: o aluno
        volta a ver apenas os posts públicos.

        **A formação continua verificada.** O caso comum é o aluno ter se
        formado ou trancado, e ele realmente estudou lá — apagar o selo
        reescreveria o passado.
      parameters:
        - { name: matriculaId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Removida }
        '401': { $ref: '#/components/responses/NaoAutenticado' }
        '403': { $ref: '#/components/responses/SomenteFaculdade' }
        '404':
          description: Matrícula não encontrada nesta instituição
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ─────────────────────────────  busca  ─────────────────────────────
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
    TipoConta:
      type: string
      enum: [aluno, faculdade, empresa]
      description: |
        Só `aluno` nasce por autocadastro. `faculdade` e `empresa` são
        provisionados à mão enquanto não há verificação institucional — e são os
        únicos tipos sem CPF.
    Universidade:
      type: object
      required: [id, nome, sigla]
      properties:
        id: { type: string, format: uuid }
        nome: { type: string, examples: ['Faculdade de Tecnologia de Ribeirão Preto'] }
        sigla: { type: string, examples: ['FATEC RP'] }
        temConta:
          type: boolean
          description: |
            Se já existe conta institucional administrando esta universidade. O
            perfil de uma sem conta não tem posts — e a tela diz isso, em vez de
            mostrar um feed vazio sem explicação.
    Curso:
      type: object
      required: [id, nome]
      properties:
        id: { type: string, format: uuid }
        nome: { type: string, examples: ['Análise e Desenvolvimento de Sistemas'] }
    Formacao:
      type: object
      required: [id, universidade, curso, verificadaEm]
      description: |
        Uma linha do currículo. `verificadaEm` não nulo é o selo — e o cliente
        mostra **apenas o selo**, sem rótulo algum nas não verificadas: a
        ausência já comunica o suficiente.
      properties:
        id: { type: string, format: uuid }
        universidade: { $ref: '#/components/schemas/Universidade' }
        curso: { $ref: '#/components/schemas/Curso' }
        verificadaEm:
          type: [string, 'null']
          format: date-time
          description: Quando a instituição confirmou. `null` = apenas declarada.
        criadoEm: { type: string, format: date-time }
    Vinculo:
      type: object
      required: [universidade, curso, criadoEm]
      description: |
        O laço ativo. **No máximo um por usuário**, e o único que concede
        visibilidade de posts `institucional` e `curso`.
      properties:
        universidade: { $ref: '#/components/schemas/Universidade' }
        curso: { $ref: '#/components/schemas/Curso' }
        criadoEm: { type: string, format: date-time }
    Matricula:
      type: object
      required: [id, cpf, curso, criadoEm, vinculada]
      description: Um aluno na lista da instituição.
      properties:
        id: { type: string, format: uuid }
        cpf:
          type: string
          description: |
            Apenas dígitos. Devolvido **só** à instituição que o cadastrou, que
            é quem o digitou — nunca em resposta pública nem ao próprio aluno de
            outra instituição.
          examples: ['12345678900']
        curso: { $ref: '#/components/schemas/Curso' }
        vinculada:
          type: boolean
          description: '`true` quando alguém já reivindicou este CPF.'
        usuario:
          description: Quem reivindicou, quando houver.
          oneOf:
            - { $ref: '#/components/schemas/PerfilPublico' }
            - { type: 'null' }
        criadoEm: { type: string, format: date-time }
    PerfilPublico:
      type: object
      required: [id, nomeCompleto, username, tipo, formacoes, criadoEm]
      properties:
        id: { type: string, format: uuid }
        nomeCompleto: { type: string }
        username: { type: string }
        tipo: { $ref: '#/components/schemas/TipoConta' }
        formacoes:
          type: array
          description: Currículo visível. Pode estar vazio.
          items: { $ref: '#/components/schemas/Formacao' }
        vinculo:
          description: |
            Instituição atual, quando há. Público de propósito: é o equivalente
            a "trabalha em" num perfil profissional.
          oneOf:
            - { $ref: '#/components/schemas/Vinculo' }
            - { type: 'null' }
        fotoUrl: { type: [string, 'null'], format: uri }
        bio: { type: [string, 'null'], maxLength: 280 }
        criadoEm: { type: string, format: date-time }
    Perfil:
      allOf:
        - $ref: '#/components/schemas/PerfilPublico'
        - type: object
          required: [email]
          description: |
            O próprio perfil. **`cpf` e `cnpj` aparecem somente aqui** — nenhum dos
            dois sai em resposta pública, nem mascarado.
          properties:
            email: { type: string, format: email }
            cpf:
              type: [string, 'null']
              description: 'Apenas dígitos. Nulo em contas faculdade e empresa.'
            cnpj:
              type: [string, 'null']
              description: |
                Apenas dígitos. Nulo em contas de aluno. Uma conta nunca tem CPF e
                CNPJ ao mesmo tempo.
            ativadaEm:
              type: [string, 'null']
              format: date-time
              description: |
                Quando a conta passou a poder agir. **Nulo significa pendente**, e
                só acontece em conta institucional: a de aluno nasce ativa.

                Pendente, a instituição entra, edita o perfil e vê a própria
                página — mas `POST` e `DELETE` em `/universidades/me/*` respondem
                `403`. O cliente usa este campo para mostrar o aviso de análise.
            telefone: { type: [string, 'null'] }
            alteradoEm: { type: [string, 'null'], format: date-time }
    PerfilDeUniversidade:
      type: object
      required: [id, nome, sigla, temVinculo, seguindo]
      properties:
        id: { type: string, format: uuid }
        nome: { type: string }
        sigla: { type: string }
        bio: { type: [string, 'null'], maxLength: 280 }
        fotoUrl: { type: [string, 'null'], format: uri }
        temVinculo:
          type: boolean
          description: |
            Se o leitor tem vínculo ativo com esta instituição. Decide se o menu
            oferece "inserir CPF" ou "encerrar vínculo".
        seguindo: { type: boolean }
        totalDeAlunos:
          type: integer
          description: Quantos têm vínculo ativo. Agregado, sem expor quem.
...
    AtualizarPerfilRequest:
      type: object
      minProperties: 1
      description: 'Envie só o que mudou. `email` e `cpf` não são editáveis.'
      properties:
        nomeCompleto: { type: string }
        username: { type: string, minLength: 3, pattern: '^[a-zA-Z0-9_]+$' }
        telefone: { type: string, pattern: '^\(?\d{2}\)?[\s-]?\d{4,5}-?\d{4}$' }
        bio: { type: [string, 'null'], maxLength: 280 }
        fotoUrl: { type: [string, 'null'], format: uri }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        fields:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
  responses:
    NaoAutenticado:
      description: Token ausente, inválido ou expirado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: nao_autenticado, message: 'Sessão expirada. Faça login novamente.' }
    SomenteFaculdade:
      description: |
        A rota exige uma conta `faculdade`. As de escrita exigem também que ela
        esteja **ativada** — uma conta pendente recebe `conta_pendente` aqui.

        A checagem de ativação consulta o banco, não um claim do token: pôr o
        estado no JWT faria uma conta desativada seguir publicando por até 15
        minutos, o tempo de vida do access token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            tipoErrado:
              value:
                code: permissao_negada
                message: 'Sua conta não tem permissão para esta ação'
            pendente:
              value:
                code: conta_pendente
                message: 'Sua instituição ainda está em análise. Você será avisado quando for ativada.'
    ValidationError:
      description: Falha de validação, com as mensagens do cliente
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            code: validation_error
            message: 'Verifique os campos destacados'
            fields:
              cpf: ['CPF inválido']
security:
  - bearerAuth: []